add: localic-std library
