+++ local-interchain/rust/.gitignore
@@ -1,1 +1,1 @@
-target/
+target
