--- local-interchain/rust/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "rust-analyzer.showUnlinkedFileNotification": false
-}
