+++ local-interchain/rust/localic-std/Cargo.toml
@@ -1,15 +1,12 @@
 [package]
-name = "localic-bin"
+name = "localic-std"
 version = "0.0.1"
 edition = "2021"
 authors = ["Strangelove Developers"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 reqwest = { version = "0.11.20", features = ["blocking", "json"]}
 tokio = { version = "1.32.0", features = ["full"] }
 serde_json = {version = "1.0.107"}
 cosmwasm-std = "1.4.0"
-
-localic-std = { path = "./localic-std"}
+thiserror = { version = "1.0.31" }
