+++ local-interchain/rust/localic-std/src/errors.rs
@@ -1,0 +1,9 @@
+use std::error;
+
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum LocalError {
+    #[error("the transaction hash was not found.")]
+    TxHashNotFound { },
+}
