+++ local-interchain/rust/localic-std/src/polling.rs
@@ -13,3 +13,5 @@ pub fn poll_for_start(c: BClient, api_url: &str, wait_seconds: u32) {
 
     panic!("Local-IC REST API Server did not start in time");
 }
+
+// TODO: polling for a future block (wait_until) delta
